requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
daphne==3.0.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mysqlclient==2.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0

txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
whitenoise==6.6.0
zope.interface==6.1
